게시판 기본 작업 완료

+++ bha-core/src/main/java/com/bupjangsa/service/BoardService.java
@@ -1,13 +1,24 @@
 package com.bupjangsa.service;
 
+import com.bupjangsa.domain.board.BoardType;
+import com.bupjangsa.domain.board.dto.BoardCriteria;
 import com.bupjangsa.domain.board.entity.Board;
 import com.bupjangsa.domain.board.infra.BoardRepository;
+import com.bupjangsa.domain.user.entity.User;
+import com.bupjangsa.domain.user.infra.UserRepository;
+import com.bupjangsa.exception.UserDataException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
+
+import static com.bupjangsa.domain.board.dto.BoardDto.*;
 
 @Service
 @RequiredArgsConstructor
@@ -15,37 +26,66 @@ import java.util.stream.Collectors;
 public class BoardService {
 
     private final BoardRepository boardRepository;
+    private final UserRepository userRepository;
 
     @Transactional
-    public void postArticle(Board board){
-        boardRepository.save(board);
+    public void postArticle(Register boardDto){
+
+        User user = userRepository.findById(boardDto.getUserId())
+                .orElseThrow(() -> new UserDataException(""));
+
+        //todo 게시판 별 번호 처리
+        Board entity = boardDto.toEntity(user);
+        boardRepository.save(entity);
     }
 
 
     @Transactional
-    public void putArticle(Board board){
-        boardRepository.findById(board.getPostId());
+    public void updateArticle(Update boardDto){
+
+        User user = userRepository.findById(boardDto.getUserId())
+                .orElseThrow(() -> new UserDataException(""));
+
+        Board board = boardRepository.findByPostNoAndBoardType(boardDto.getPostNo(), boardDto.getBoardType())
+                .orElseThrow(() -> new RuntimeException(""));
+
+        board.updateBoardData(boardDto.getTitle(), boardDto.getContents(), user);
     }
 
     @Transactional
-    public void deleteArticle(Board board){
+    public void deleteArticle(Delete boardDto){
+
+        User user = userRepository.findById(boardDto.getUserId())
+                .orElseThrow(() -> new UserDataException(""));
+
+        Board board = boardRepository.findByPostNoAndBoardType(boardDto.getPostNo(), boardDto.getBoardType())
+                .orElseThrow(() -> new RuntimeException(""));
+
+        //등록자가 아닐 경우 예외처리
+        if(!board.getCreatedBy().getAccountId().equals(user.getAccountId())){
+            throw new RuntimeException("");
+        }
+
         boardRepository.delete(board);
     }
 
     //단건 조회
-    public Board selectArticle(String boardType, int boardNo){
-        return boardRepository.selectArticle(boardType, boardNo);
+    public PostInfo selectArticle(BoardType boardType, Long postNo){
+        return boardRepository.findByPostNoAndBoardType(postNo, boardType)
+                .map(PostInfo::from)
+                .orElseThrow(() -> new RuntimeException(""));
     }
 
     //게시물 목록 조회
-    public List<Board> selectArticleList(String boardType, int pageNo, int pageSize){
-        // 리스트 페이징. TODO 게시물이 많아질때 처리 고민 필요.
-        List<Board> boardList = boardRepository.selectArticleList(boardType);
+    public Page<PostInfo> selectArticleList(BoardCriteria.SearchList criteria, Pageable pageable){
 
-        return boardList.stream()
-                .skip(pageSize * (pageNo -1))
-                .limit(pageSize)
+        Page<Board> boardPage = boardRepository.selectArticlePage(criteria, pageable);
+
+        final var boardInfoList = boardPage.getContent().stream()
+                .map(PostInfo::from)
                 .collect(Collectors.toList());
+
+        return new PageImpl<>(boardInfoList, pageable, boardPage.getTotalElements());
     }
 
 }
